Upgrade CLI hierarchy
