+++ src/main/java/com/excilys/cdb/main/Main.java
@@ -1,12 +1,12 @@
 package com.excilys.cdb.main;
 
-import com.excilys.cdb.ui.UseMenu;
+import com.excilys.cdb.ui.CLIMenu;
 
 public class Main {
 
 	public static void main(String[] args) {
 
-		new UseMenu().useMenu();
+		new CLIMenu().useMenu();
 
 	}
 }
