+++ src/main/java/com/excilys/cdb/ui/Menu.java
@@ -14,37 +14,4 @@ public class Menu {
 		System.out.println("| 3.         Quit          |");
 		System.out.println("+--------------------------+");
 	}
-	
-	public void afficherMenuComputer() {
-		
-		System.out.println("+--------------------------+");
-		System.out.println("|       Computer Menu      |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 1.   Create a Computer   |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 2.   Delete a Computer   |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 3.   Update a Computer   |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 4.    Find a Computer    |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 5.   List all Computers  |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 6.         Quit          |");
-		System.out.println("+--------------------------+");
-	}
-	
-	public void afficherMenuCompany() {
-		
-		System.out.println("+--------------------------+");
-		System.out.println("|       Company Menu       |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 1.    Find a Company     |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 2.   List all Companies  |");
-		System.out.println("+--------------------------+");
-		System.out.println("| 3.         Quit          |");
-		System.out.println("+--------------------------+");
-	}
-
 }
